Bug fix: create_or_delete token was completely broken
Bug one: fixed a call to RoomData.game_state that was missing the
game_state part.

Bug two: I thought popping the positions off of the id_to_positions
dictionary was more clear than just overwriting it later, but the code
was instead popping one position off of the position list, and crashing
when updating tokens. Changed it to simply reading the value, and
overwriting it after.