Hand roll dict to room entities
In rooms with lots of tokens, we were spending ~30% of the update
letting dacite handle this for us. Hand-rolling it makes it <1%.

The real long-term fix is to move to incremental updates so we're not
encoding/decoding the whole room every update, but this is a good
stop-gap to get better performance.