Fix firefox search tray drag
Firefox does not like it when you remove the target element inside of the pointerdown event, it will stop giving any pointer events for the entirety of the drag, regardless of what element the event handlers are attached to. This doesn't happen if we wait until the next event loop run to trigger the removal.